+++ src/main.ts
@@ -13,8 +13,7 @@ import { SRSettingTab, SRSettings, DEFAULT_SETTINGS } from "src/settings";
 import { FlashcardModal, Deck } from "src/flashcard-modal";
 import { StatsModal, Stats } from "src/stats-modal";
 import { ReviewQueueListView, REVIEW_QUEUE_VIEW_TYPE } from "src/sidebar";
-import { Card, ReviewResponse, schedule } from "src/scheduling";
-import { CardType } from "src/types";
+import { Card, CardType, ReviewResponse, schedule } from "src/scheduling";
 import {
     YAML_FRONT_MATTER_REGEX,
     SCHEDULING_INFO_REGEX,
