+++ src/scheduling.ts
@@ -1,7 +1,6 @@
 import { TFile } from "obsidian";
 
 import { SRSettings } from "src/settings";
-import { CardType } from "src/types";
 import { t } from "src/lang/helpers";
 
 export enum ReviewResponse {
@@ -32,6 +31,14 @@ export interface Card {
     // information for sibling cards
     siblingIdx: number;
     siblings: Card[];
+}
+
+export enum CardType {
+    SingleLineBasic,
+    SingleLineReversed,
+    MultiLineBasic,
+    MultiLineReversed,
+    Cloze,
 }
 
 export function schedule(
