--- src/types.ts
@@ -1,11 +1,0 @@
-// https://github.com/obsidianmd/obsidian-api/issues/13
-
-// flashcards
-
-export enum CardType {
-    SingleLineBasic,
-    SingleLineReversed,
-    MultiLineBasic,
-    MultiLineReversed,
-    Cloze,
-}
