+++ tsconfig.json
@@ -14,6 +14,7 @@
         "esModuleInterop": true,
         "importHelpers": true,
         "lib": ["dom", "dom.Iterable", "scripthost", "ES5", "ES6", "ES7", "ES2021"],
+        "jsx": "react",
         "jsxFactory": "h",
     },
     "include": ["**/*.ts", "**/*.tsx"]
